feat(VaultClient): construct a new `VaultClient` with no lookup
The standard function, `VaultClient::new`, for constructing a
`VaultClient` makes a call to the `auth/token/lookup-self` endpoint to
populate additional information about the token. However, for limited
use tokens, you will not want to perform this lookup.

As a result, the `VaultClient.data` has now been made an `Option`.

[breaking-change]